Добавлен перечень ресурсов. Доработаны сервисы безопасности. Добавлено отображение количества отображенных записей из общего числа

+++ RequestsForRightsV2/Controllers/ResourceGroupController.cs
@@ -1,8 +1,9 @@
 ﻿using System;
 using System.Data.Entity.Infrastructure;
-using System.Net.Http;
 using System.Web.Mvc;
 using RequestsForRights.Domain.Entities;
+using RequestsForRightsV2.Infrastructure.Helpers;
+using RequestsForRightsV2.Infrastructure.Security.Interfaces;
 using RequestsForRightsV2.Infrastructure.Services.Interfaces;
 using RequestsForRightsV2.Models.FilterOptions;
 
@@ -11,36 +12,61 @@ namespace RequestsForRightsV2.Controllers
     public class ResourceGroupController : Controller
     {
         private readonly IResourceGroupService _resourceGroupService;
-        
-        public ResourceGroupController(IResourceGroupService resourceGroupService)
+        private readonly IResourceGroupSecurityService _securityService;
+
+        public ResourceGroupController(IResourceGroupService resourceGroupService,
+            IResourceGroupSecurityService securityService)
         {
             if (resourceGroupService == null)
             {
                 throw new ArgumentNullException("resourceGroupService");
             }
             _resourceGroupService = resourceGroupService;
+            if (securityService == null)
+            {
+                throw new ArgumentNullException("securityService");
+            }
+            _securityService = securityService;
         }
-        //
-        // GET: /ResourceGroup/
+
         public ActionResult Index(FilterOptions filterOptions)
         {
+            if (!_securityService.CanRead())
+            {
+                return RedirectToAction("ForbiddenError", "Home");
+            }
+            ViewData["SecurityService"] = _securityService;
             return View(_resourceGroupService.GetResourceGroupIndexModelView(filterOptions));
         }
 
-        public PartialViewResult GetDataTable(FilterOptions filterOptions)
+        public ActionResult GetDataTable(FilterOptions filterOptions)
         {
+            if (!_securityService.CanRead())
+            {
+                return RedirectToAction("ForbiddenError", "Home");
+            }
+            ViewData["SecurityService"] = _securityService;
             return PartialView("DataTable", _resourceGroupService.GetResourceGroupIndexModelView(filterOptions));
         }
 
         [HttpGet]
         public ActionResult Update(int id)
         {
+            if (!_securityService.CanUpdate())
+            {
+                return RedirectToAction("ForbiddenError", "Home");
+            }
+            ViewData["SecurityService"] = _securityService;
             return View(_resourceGroupService.GetResourceGroupById(id));
         }
 
-        [HttpPost]
+        [HttpPut]
         public ActionResult Update(ResourceGroup resourceGroup)
         {
+            if (!_securityService.CanUpdate(resourceGroup))
+            {
+                return RedirectToAction("ForbiddenError", "Home");
+            }
             if (resourceGroup == null)
             {
                 Response.StatusCode = 400;
@@ -57,8 +83,8 @@ namespace RequestsForRightsV2.Controllers
             }
             catch (DbUpdateException e)
             {
-                Response.StatusCode = 409;
-                return Content(e.Message);
+                return RedirectToAction("ConflictError", "Home",
+                    new { message = ExceptionHelper.RollToInnerException(e).Message });
             }
             return Request["returnUri"] != null ? (ActionResult)Redirect(Request["returnUri"]) : 
                 RedirectToAction("Index");
@@ -66,12 +92,22 @@ namespace RequestsForRightsV2.Controllers
 
         public ActionResult Detail(int id)
         {
+            if (!_securityService.CanRead())
+            {
+                return RedirectToAction("ForbiddenError", "Home");
+            }
+            ViewData["SecurityService"] = _securityService;
             return View(_resourceGroupService.GetResourceGroupById(id));
         }
 
         [HttpDelete]
         public ActionResult Delete(int id)
         {
+            var resourceGroup = _resourceGroupService.GetResourceGroupById(id);
+            if (!_securityService.CanDelete(resourceGroup))
+            {
+                return RedirectToAction("ForbiddenError", "Home");
+            }
             try
             {
                 _resourceGroupService.DeleteResourceGroup(id);
@@ -79,8 +115,8 @@ namespace RequestsForRightsV2.Controllers
             }
             catch (DbUpdateException e)
             {
-                Response.StatusCode = 409;
-                return Content(e.Message);
+                return RedirectToAction("ConflictError", "Home",
+                    new { message = ExceptionHelper.RollToInnerException(e).Message });
             }
             return RedirectToAction("GetDataTable");
         }
@@ -88,12 +124,21 @@ namespace RequestsForRightsV2.Controllers
         [HttpGet]
         public ActionResult Create()
         {
+            if (!_securityService.CanCreate())
+            {
+                return RedirectToAction("ForbiddenError", "Home");
+            }
+            ViewData["SecurityService"] = _securityService;
             return View();    
         }
 
         [HttpPost]
         public ActionResult Create(ResourceGroup resourceGroup)
         {
+            if (!_securityService.CanCreate(resourceGroup))
+            {
+                return RedirectToAction("ForbiddenError", "Home");
+            }
             if (resourceGroup == null)
             {
                 Response.StatusCode = 400;
@@ -110,8 +155,8 @@ namespace RequestsForRightsV2.Controllers
             }
             catch (DbUpdateException e)
             {
-                Response.StatusCode = 409;
-                return Content(e.Message);
+                return RedirectToAction("ConflictError", "Home",
+                    new { message = ExceptionHelper.RollToInnerException(e).Message });
             }
             return RedirectToAction("Index");
         }
